feat: created CookieBanner component
This component is rendered at the bottom of the content area, and is on the top layer (z-index). In case the user accepts the essential cookies, a new cookie will be created that indicates that a user has accepted the cookies. In case the user rejects the cookies all the existing cookies are removed. The removal of cookies may be used in the future to allow the user to change his cookie-preference.