feat: created useForm hook
This hook is used to update an multi-layered object in the context of a form. Each input in the form must have a name property that matches the properties' name that it represents. In case the property is below the top-layer, the data-path property must be set to the property tree in the form of an array. For example ['station', 'address']. The tree of properties given with the `path` property is then combined with the name attribute to change the properties value.

In order to match the properties type, the `data-result` property is used. It specifies the type of the properties' value.   The `data-result` property can be set to the following values:
- string
- number
- array<number>
- array<string>