feat: created RenderStation component
This component renders a single Station. It allows users to edit that station and then update it or just to delete that specific station. The editable property of this component either enables or disables the edit and delete functionality, so that this component can be used both for regular and management users.